HIP-940: Fix type conversion

+++ src/ska-sdp-func/visibility/sdp_opt_weighting.cpp
@@ -928,7 +928,7 @@ void sdp_optimized_weighting(
 
         int n_threads  = tile_size_u * tile_size_v;
         const uint64_t num_threads_briggs[] = {n_threads, 1, 1};
-        const uint64_t num_blocks_briggs[] = {(int)num_tiles - 1, 1, 1};
+        const uint64_t num_blocks_briggs[] = {(uint64_t)num_tiles - 1, 1, 1};
 
         // Make sure weights are on the gpu
         sdp_mem_check_location(output_weights, SDP_MEM_GPU, status);
@@ -1081,7 +1081,7 @@ void sdp_optimised_indexed_weighting(
 
         int n_threads  = tile_size_u * tile_size_v;
         const uint64_t num_threads_briggs[] = {n_threads, 1, 1};
-        const uint64_t num_blocks_briggs[] = {(int)num_tiles - 1, 1, 1};
+        const uint64_t num_blocks_briggs[] = {(uint64_t)num_tiles - 1, 1, 1};
 
         // Make sure weights are on the gpu
         sdp_mem_check_location(output_weights, SDP_MEM_GPU, status);
